Claude/fix nextjs blog styling 012zmmp dka9 hoa gyb9 ha bpi b (#173)
* Fix Tailwind CSS v4 configuration and update documentation

- Add missing @tailwindcss/postcss and postcss dependencies required for Tailwind v4
- Create postcss.config.mjs with proper Tailwind v4 plugin configuration
- Update README with Tailwind v4 information and build requirements
- Document need for Drupal API access during static build

This fixes the styling issues where Tailwind CSS classes weren't being processed
correctly. Tailwind v4 requires the @tailwindcss/postcss plugin to work with
Next.js, which was missing from the project.

* Add @tailwindcss/typography plugin import for Tailwind v4

In Tailwind CSS v4, plugins must be imported using the @plugin directive in CSS
files rather than configured in JavaScript. This fixes the "Cannot apply unknown
utility class 'prose'" error.

* Fix routing by updating to App Router compatible API methods

The code was using Pages Router API methods (getResourceFromContext,
getResourceCollectionFromContext) which don't work correctly with Next.js
App Router. Updated all routes to use the correct methods:

- Blog posts: Use translatePath() + getResource() instead of getResourceFromContext()
- Collections: Use getResourceCollection() instead of getResourceCollectionFromContext()
- Updated all generateStaticParams() and metadata generation functions

This fixes the 404 errors when accessing blog posts and other dynamic routes.

* Fix TypeScript error: use correct DrupalTranslatedPath properties

The translatePath() method returns an object with nested entity properties.
Updated to use resource.entity.type and resource.entity.uuid instead of
resource.type and resource.id.

* Add detailed logging to generateStaticParams for debugging

Added comprehensive console logging to help diagnose why static pages aren't
being generated. This will show:
- When the API is being called
- How many posts are found
- Each path being generated
- Any errors that occur

IMPORTANT: When generateStaticParams returns an empty array (e.g., API failure),
Next.js will NOT generate any static HTML files for those routes, causing 404s.

* Fix 404s by using getResourceByPath instead of translatePath

The translatePath() method was failing during build with 404 errors because
the /router/translate-path Drupal endpoint wasn't working correctly.

Switched to getResourceByPath() which:
- Fetches the resource directly by path in one API call
- Avoids the problematic translate-path endpoint
- Is more efficient and reliable

This should resolve the 45+ 404 errors seen during static generation.

* Add Vercel configuration for static export

Configure Vercel to properly serve the static export:
- Set output directory to 'out'
- Enable trailing slashes to match Next.js config
- Disable clean URLs to preserve directory structure

This fixes 404 errors on Vercel preview deployments.

* Revert "Add Vercel configuration for static export"

This reverts commit 96b09aa75257b9511bc5b13a50b182eead7c833a.

* Update test.yml

* Update test.yml

* Add node__title class for Behat test compatibility

Added .node__title class to BlogPostCard h2 element to maintain
compatibility with existing Behat tests that expect this selector.

* Update page.tsx

* Migrate from App Router to Pages Router for battle-tested static export

This migration switches from Next.js App Router to Pages Router using
getStaticPaths and getStaticProps, which provides 100% reliable static
page generation for all deployment platforms (Netlify, Vercel, S3, etc.).

Changes:
- Created src/pages/_app.tsx with layout and Tailwind CSS imports
- Migrated all routes to Pages Router:
  - src/pages/index.tsx (home page with blog listing)
  - src/pages/[...slug].tsx (individual blog posts)
  - src/pages/blog/[page].tsx (pagination)
  - src/pages/tag/[tid].tsx (tag pages)
  - src/pages/about.tsx (about page)
  - src/pages/talks.tsx (talks page)
  - src/pages/404.tsx (custom 404 page)
- Removed old App Router page files
- Converted metadata handling from App Router exports to Head component
- Created static robots.txt and manifest.json in public directory
- Updated .gitignore to allow src/pages (was excluded for old Gatsby)

All routes use proper TypeScript typing with GetStaticPaths and
GetStaticProps for type-safe static generation.

* Rename globals.css to globals.css

* Delete src/app directory

* Update _app.tsx

---------
