+++ .github/workflows/test.yml
@@ -57,7 +57,7 @@ jobs:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
 
       - name: Start the serving of the site
-        run: npx --yes serve@latest -l http://0.0.0.0 -p 9000 out
+        run: npx --yes serve@latest -l tcp://0.0.0.0:9000 out &
 
       - name: Make sure serving has started
         run: ./vendor/bin/wait-for-listen 9000
