--- src/app/not-found.tsx
@@ -1,19 +1,0 @@
-import Link from "next/link"
-
-export default function NotFound() {
-  return (
-    <div className="max-w-4xl mx-auto text-center py-16">
-      <h1 className="text-6xl font-bold text-gray-900 mb-4">404</h1>
-      <h2 className="text-3xl font-semibold text-gray-700 mb-6">Page Not Found</h2>
-      <p className="text-gray-600 mb-8">
-        Sorry, the page you're looking for doesn't exist or has been moved.
-      </p>
-      <Link
-        href="/"
-        className="inline-block bg-blue-600 text-white px-6 py-3 rounded-lg hover:bg-blue-700 transition-colors"
-      >
-        Go Back Home
-      </Link>
-    </div>
-  )
-}
