--- src/app/robots.ts
@@ -1,13 +1,0 @@
-import { MetadataRoute } from 'next'
-
-export const dynamic = 'force-static'
-
-export default function robots(): MetadataRoute.Robots {
-  return {
-    rules: {
-      userAgent: '*',
-      allow: '/',
-    },
-    sitemap: 'https://eiriksm.dev/sitemap.xml',
-  }
-}
