--- src/app/sitemap.ts.disabled
@@ -1,58 +1,0 @@
-import { MetadataRoute } from 'next'
-import { drupal } from '@/lib/drupal'
-import { DrupalNode } from 'next-drupal'
-import { getNodePath } from '@/lib/utils'
-
-export const revalidate = 3600 // Revalidate every hour
-
-export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
-  const baseUrl = process.env.NEXT_PUBLIC_SITE_URL || 'https://eiriksm.dev'
-
-  let articles: MetadataRoute.Sitemap = []
-
-  try {
-    // Fetch all articles
-    const nodes = await drupal.getResourceCollectionFromContext<DrupalNode>(
-      'node--article',
-      {
-        params: {
-          'sort': '-created',
-        },
-      }
-    )
-
-    // Generate sitemap entries for articles
-    articles = nodes.map((node) => ({
-      url: `${baseUrl}${getNodePath(node)}`,
-      lastModified: new Date(node.changed || node.created),
-      changeFrequency: 'monthly' as const,
-      priority: 0.8,
-    }))
-  } catch (error) {
-    console.error('Failed to fetch articles for sitemap:', error)
-  }
-
-  // Static pages
-  const staticPages = [
-    {
-      url: baseUrl,
-      lastModified: new Date(),
-      changeFrequency: 'daily' as const,
-      priority: 1,
-    },
-    {
-      url: `${baseUrl}/about`,
-      lastModified: new Date(),
-      changeFrequency: 'monthly' as const,
-      priority: 0.5,
-    },
-    {
-      url: `${baseUrl}/talks`,
-      lastModified: new Date(),
-      changeFrequency: 'monthly' as const,
-      priority: 0.5,
-    },
-  ]
-
-  return [...staticPages, ...articles]
-}
